requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
httpx==0.27.2
python-dotenv==1.0.1
jinja2==3.1.4
pydantic-settings==2.6.1
python-multipart==0.0.12
anthropic>=0.40.0
Pillow>=10.0.0
pycognito>=0.4.0
qdrant-client>=1.9.0
fastembed>=0.3.0
gspread>=6.0.0
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
